requirements.txt
Django==3.0.2
django-import-export==2.0.1
